requirements.txt
streamlit
pandas
matplotlib
python-dotenv
requests
apify-client   
textblob
google-generativeai